migrated code from APP to separate page
