+++ src/WhatLearnt.txt
@@ -1,2 +1,3 @@
 1) how to create a reusable Card component, fetch data from an API, and display multiple results in these cards using React
 2)Skeleton loading with infinite scrolling
+3)Migrate the whole infinite sccroll and skeleton loading code to a component and fix routes
